refactor(plugin): simplify run_mobile_plugin return types
