+++ core/tauri/src/window.rs
@@ -1530,8 +1530,7 @@ impl<R: Runtime> Window<R> {
                     plugin: &str,
                     runtime_handle: &R::Handle,
                     message: InvokeMessage<R>,
-                    callback: CallbackFn,
-                    error: CallbackFn,
+                    (callback, error): (CallbackFn, CallbackFn),
                     env: JNIEnv<'_>,
                     activity: JObject<'_>,
                     webview: JObject<'_>,
@@ -1554,7 +1553,7 @@ impl<R: Runtime> Window<R> {
                         webview.into(),
                         env.new_string(plugin)?.into(),
                         env.new_string(&message.command)?.into(),
-                        data.into(),
+                        data,
                         (callback.0 as i64).into(),
                         (error.0 as i64).into(),
                       ],
@@ -1567,8 +1566,7 @@ impl<R: Runtime> Window<R> {
                     &plugin,
                     &runtime_handle,
                     message,
-                    resolver.callback,
-                    resolver.error,
+                    (resolver.callback, resolver.error),
                     env,
                     activity,
                     webview,
